+++ src/lib/output/themes/default/assets/typedoc/utils/decompress.ts
@@ -1,7 +1,7 @@
 /**
- * Decompresses Base64-encoded Gzip data and parses it into a JSON object.
+ * Decompresses Base64-encoded deflate compressed data and parses it into a JSON object.
  *
- * @param base64 - The Base64-encoded string representing the Gzip-compressed JSON string.
+ * @param base64 - The Base64-encoded string representing the deflate-compressed JSON string.
  * @returns A promise that resolves to the parsed JSON object.
  */
 export async function decompressJson(base64: string) {
@@ -9,7 +9,7 @@ export async function decompressJson(base64: string) {
     const blob = new Blob([binaryData]);
     const decompressedStream = blob
         .stream()
-        .pipeThrough(new DecompressionStream("gzip"));
+        .pipeThrough(new DecompressionStream("deflate"));
     const decompressedText = await new Response(decompressedStream).text();
     return JSON.parse(decompressedText);
 }