+++ CHANGELOG.md
@@ -13,6 +13,7 @@ title: Changelog
 -   Switch from gzip to deflate for compressing assets to make output consistent across different operating systems, #2796.
 -   `@include` and `@includeCode` now work for comments on the entry point for projects with a single entry point, #2800.
 -   Cascaded modifier tags will no longer be copied into type literals, #2802.
+-   `@summary` now works to describe functions within modules, #2803.
 
 ## v0.27.3 (2024-12-04)